+++ sidplay/sidplay2.cpp
@@ -119,7 +119,7 @@ void sidplay2::stop (void)
 void sidplay2::pause (void)
 {   sidplayer.pause (); }
 
-uint_least32_t sidplay2::play (float *buffer, uint_least32_t count)
+uint_least32_t sidplay2::play (short *buffer, uint_least32_t count)
 {   return sidplayer.play (buffer, count); }
 
 int sidplay2::load (SidTune *tune)
